+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   ones that have a 'NotAvailableForSubscription'-restriction for the location.
 - Taking VM generation into account when selecting an Azure role size / image.
 - Added parameter 'UseAzureUrl' to 'Get-LabTfsFeed'.
+- Added parameter 'UseAzureUrl' to 'Install-LabBuildWorker' because of unexpected results in the DscWorkshop.
 - Add *.azure.com to Local Intranet Zone in Azure VM initialization.
 
 ### Bugs